Refactoring the footer (#282)
Fix #205 

Use HorizontalFlexContainer so it is much cleaner.

I also add "What is a hardware wallet?" which was previously disabled.

# Big screen

<img width="1572" alt="screen shot 2019-02-06 at 1 38 51 am" src="https://user-images.githubusercontent.com/31379986/52346196-6a4b6b80-2a62-11e9-96d8-1d6d51de3dce.png">

# Small screen

<img width="918" alt="screen shot 2019-02-06 at 1 39 30 am" src="https://user-images.githubusercontent.com/31379986/52346203-6cadc580-2a62-11e9-9618-9c02fc6fc7dd.png">
