+++ app/components/layout/HorizontalFlexContainer.js
@@ -1,0 +1,21 @@
+// @flow
+import React, { Component } from 'react';
+import type { Node } from 'react';
+import { observer } from 'mobx-react';
+import styles from './HorizontalFlexContainer.scss';
+
+type Props = {
+  children: ?Node,
+};
+
+@observer
+export default class HorizontalFlexContainer extends Component<Props> {
+  render() {
+    const { children } = this.props;
+    return (
+      <div className={styles.component}>
+        {children}
+      </div>
+    );
+  }
+}
