+++ app/components/layout/HorizontalFlexContainer.scss
@@ -1,0 +1,12 @@
+.component {
+  width: 100%;
+  display: flex;
+  flex-direction: row;
+  overflow-x: overlay;
+  position: relative;
+
+  &::-webkit-scrollbar-button {
+    width: 7px;
+    display: block;
+  }
+}
